Seperate the code into View, ViewModel, Domain Layer and Model Layer
Mainly Presentation logic are put into the view Model.

the UI layer can also contain business logic. This can be the case when combining data from multiple repositories to create the screen UI state, or when a particular type of data doesn't require a data layer.

ViewModel is the right place to handle business logic in the UI layer. The ViewModel is also in charge of handling events and delegating them to other layers of the hierarchy when business logic needs to be applied to modify application data.

This is the functionalties of a domain layer. I call those UIDomain